* WavePlayer now always plays an empty sound file at startup, making sure that the entire sound output chain is properly initialized prior to playing the first actual sound. * Simplified writing of CMakeLists by automatically dealing with headers, includes, and solution folders. * Added BLOB data type. * Added infrastructure for binary resources. * Created CMake infrastructure for automated testing (not functional yet). * BCIExceptions may now inherit from other std::exception descendants.
git-svn-id: http://www.bci2000.org/svn/trunk@4447 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/SignalSource/NicoletOne/LiveData.cpp
@@ -96,7 +96,7 @@ void CLiveData::_GrabData()
          {
             // 's' is the datagram from channel 'liDataSerie' and sample 'liSample' in segment 'liDataSegment'
             short s = *pData++;
-		    mCurData[liDataSerie].push( ( double )s * dResolution + dOffset );
+    		    mCurData[liDataSerie].push( static_cast<float>( s * dResolution + dOffset ) );
          }
 			 
       }
@@ -138,7 +138,7 @@ bool CLiveData::GetData( int ChannelSize, int DataBlockSize, float** DataBlock, 
 		return true;
 	}
 	// Return false if we don't have any data for this channel
-	if( mCurData[0].size() < DataBlockSize )
+	if( static_cast<int>( mCurData[0].size() ) < DataBlockSize )
 	{
 		mDataLock->Unlock();
 		return true;
